requirements.txt
# Web Framework
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
pydantic>=2.4.2

# Google Cloud Services
google-auth==2.23.3
google-cloud-bigquery==3.11.4
google-cloud-storage==2.13.0
google-cloud-aiplatform>=1.35.0


# Data Processing & Analysis
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
scipy==1.11.3

# Visualization & Image Processing
matplotlib==3.8.0
wordcloud==1.9.2
Pillow==10.0.1

# Document Generation
python-pptx==0.6.21
python-docx==1.0.1
reportlab==4.0.4
PyPDF2==3.0.1

# NLP & Text Processing
regex==2023.8.8
nltk==3.8.1

# Email Services
python-jose==3.3.0
emails==0.6
secure-smtplib==0.1.1

# Database & Search
elasticsearch==8.11.1  # Added for Elasticsearch integration
db-dtypes==1.1.1      # For BigQuery data types

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
tqdm==4.66.1
pytz==2023.3.post1
pathlib==1.0.1        # Added for path handling
shutil-which==3.0.0   # Added for file operations

# Testing & Development
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.23.5  # For testing async code
black==24.1.1          # Code formatting
flake8==7.0.0          # Code linting
